+++ .vscode/settings.json
@@ -1,0 +1,12 @@
+{
+	"files.associations": {
+		"*.c": "c",
+		"cub3d.h": "c",
+		"utils.h": "c",
+		"stdlib.h": "c",
+		"libft.h": "c",
+		"render.h": "c",
+		"control.h": "c",
+		"parse.h": "c"
+	}
+}
