+++ Makefile
@@ -1,7 +1,7 @@
 NAME		:= cub3D
 LIBFT_DIR	:= ./libft
 LIBFT		:= $(LIBFT_DIR)/libft.a
-MLX_DIR		:= ./minilibx
+MLX_DIR		:= ./minilibx-linux
 MLX_LIB		:= $(MLX_DIR)/libmlx.a
 INCS_DIR	:= ./include
 SRCS_DIR	:= ./src
@@ -10,13 +10,13 @@ OBJS_DIR	:= ./objs
 CC			:= cc
 CCFLAGS		:= -Wall -Wextra -Werror -I $(INCS_DIR) -I $(MLX_DIR) -I $(LIBFT_DIR) -g
 # LDFLAGS		:= -lm -L/usr/X11/lib -lX11 -lXext
-# macOS specific flags
-LDFLAGS		:= -L$(MLX_DIR) -lmlx -framework OpenGL -framework AppKit
-MLXFLAGS	:= -L$(MLX_DIR) -lmlx -framework OpenGL -framework AppKit
+MLXFLAGS	:= -L$(MLX_DIR) -lmlx -lXext -lX11 -lm -lz
 
 SRCS		:= $(SRCS_DIR)/main.c \
 			   $(SRCS_DIR)/control/input.c \
+			   $(SRCS_DIR)/control/input_2.c \
 			   $(SRCS_DIR)/parse/init_game.c \
+			   $(SRCS_DIR)/parse/init_game_2.c \
 			   $(SRCS_DIR)/parse/load_map.c \
 			   $(SRCS_DIR)/parse/map_check.c \
 			   $(SRCS_DIR)/parse/color.c \
@@ -24,15 +24,15 @@ SRCS		:= $(SRCS_DIR)/main.c \
 			   $(SRCS_DIR)/parse/file_io.c \
 			   $(SRCS_DIR)/parse/classification.c \
 			   $(SRCS_DIR)/parse/map_builder.c \
+			   $(SRCS_DIR)/parse/map_builder_2.c \
 			   $(SRCS_DIR)/parse/config.c \
 			   $(SRCS_DIR)/render/raycast.c \
 			   $(SRCS_DIR)/render/init_render.c \
+			   $(SRCS_DIR)/render/init_render_2.c \
 			   $(SRCS_DIR)/render/texture.c \
 			   $(SRCS_DIR)/render/calc.c \
 			   $(SRCS_DIR)/render/draw.c \
 			   $(SRCS_DIR)/render/ray.c \
-			   $(SRCS_DIR)/util/debug.c \
-			   $(SRCS_DIR)/util/debug_display_textures.c \
 			   $(SRCS_DIR)/util/error_exit.c \
 			   $(SRCS_DIR)/util/free_game.c \
 			   $(SRCS_DIR)/util/free_init_resources.c \
@@ -51,7 +51,7 @@ $(MLX_LIB):
 	$(MAKE) -C $(MLX_DIR)
 
 $(NAME): $(OBJS) $(LIBFT) $(MLX_LIB)
-	$(CC) $(CCFLAGS) -o $@ $^ $(LDFLAGS)
+	$(CC) $(CCFLAGS) -o $@ $^ $(MLXFLAGS)
 
 $(OBJS_DIR)/%.o: $(SRCS_DIR)/%.c
 	@mkdir -p $(dir $@)
@@ -66,5 +66,8 @@ fclean: clean
 	rm -f $(NAME) $(LIBFT) $(MLX_LIB)
 
 re: fclean all
+
+rr:
+	make re && make clean
 
 .PHONY: all clean fclean re
